Make inline memory initialization the default
Fixes #1293
